requirements.txt
Flask
Redis
requests
mysql-connector
